Move get_repo_name() to the output module
